Nest extracts from different frames
Instead of storing the extracted frame number as an array these changed
nest the frame as the top level key of the retued object.  This will
ensure that all classifications made one the same frame are extracted to
the same lists and making it easier for the front-end to grab the data
for the shown frame.