Add the ability to auto detect a text subtask on a polygon tool
This detection is quite ugly at the moment and only works for the case
of a test in a polygon dring tool.  A more generalized subtask detection
should be created at a later date.  At that time the hardcoded text tool
extractor in the `poly_line_text_extractor` should also be externalized.