Update csv_utils to allow poly-line-text reducer to run offline
This update also required the surevey extractor/reducer pair to be
updated to used `_` instead of `.` in the key names.

This also reveiled a bug that would have prevented the online version of
survye reduction from running.  The reducer was expecting a flatter data
structure than it would have received from the extractor.  The tests and
code have been updated to reflect the true form of this data.