Create and use a decorator for extractros
To reduce the ammount of repeated code in the extractor definitions a
decorator is now being used.  The decorator will automatically detect if
a flask `reqest` or a `obejct` is being pass in and change the function
call accordingly.

When a `request` is passed in the decorator will also transform the
annotations list back to the "raw" form (ceasar changes the shape of the
data before passing to extractros) so it can be passed into this code
without error.