Create a decorator for reducer functions
This abstracts aways a large chunk of request framework of reducers as a
decorator.  When a flask `request` is passed in it will correctly
extract the data out of the request.  In all other cases the data is
processed as is.

It also has the ability to pass in the `process_data` function and the
`DEFAULT` keywords as arguments to the decorator.

A reference to the orginal functon (the one being wrapped) is kept to
allow for better testing and to allow for other reducers to used it.