Update reduction to include consensus score and better word list style
Word lists now include empty strings when a user's transcription does
not include a word.  This allows the orginal text strings to be
reconstructed from the reduction data.

Two new fields have been added to each line's reduction, `number_views`
and `consensus_score`.  This are the number of users that have seen the
line of text and the average agreement of words in the line. If the two
value are the same every user agrees with each other.