Create decorator to correctly extract subtasks
The extractor needs to have the `details` keyword passed in.  This
keyword is a dictionary with one key per tool with subtask(s), formated
as `T0_tool0`, `T1_tool1`, etc...  The values are a list of extractors
to used for each subtask(s).  This dictionary can be passed in as a URL
parameter as well.