Change expected annotation payload for offline extractors
This makes the annotation format of Caesar the expected form for offline 
extraction.  The base test class has been updated to handle this 
conversion.  The added bonus of doing this this way is the same keywords 
are now passed in both modes.

Also the `filter_annotations` function is no longer needed and will be 
removed in a later commit.