Add aggregation version number to all extractors
This is so we can track what version of the code produeced each extact.  
The version appending is done in the wrapper so it is applied the same 
way to all extractors.

The test base class was also updated so none of the individual tests 
need to include this information.