Add aggregation version to reductions
The `append_version` code was extracted in its own file so it can be 
used for both the extractors and the reducers.

The version number of the aggregtion code is now included in each 
reduction.  This has been done at the wrapper level so it is applied to 
all reducers in the same way.

The base test class has been updated to add this where it is expected, 
that way none of the individual tests need to be updated.