Updated reducer auto configuation to include subtask keywords
The poly_line_text_reducer will also have the `dot_freq` keyword pass 
along into it.

The reducer config function has been moved out of the command line 
script and into the same file as the extracotr config function.  A new 
test was added to ensure it works as expected.