Add try-except case to setup.py when checking for README.md file
The doceker cotainer "builds" the module before moving the README.md 
file to save time on not re-building when they do not change.  This 
caused a crash with the latest change that added the README.md as the 
`long_description`.  Now setup.py uses a `try` block to catch this 
error.