Add shape reducer using DBSCAN
This is the reducer to be used with the shape extractor.  The `shape` 
keyword must be provided.  Any shape with and `angle` paramter is 
restricted to using a custom `euclidean` metric that assumes the last 
column of the data array contains the angles.  A special `mean` function 
for angles is also included that makes the same assumption about the 
order of paramters.

Only a few tests have been added to start, more to come.