Add ablility to normalize symetric shapes
Some shapes are symetric and the clustering should take this into 
account.  This commit generalizes the various angle utility functions to 
account for these symetries when the `symetric` keyword is used.

The normaization is done such that the retured shape has the same 
paramters as the non-normaized shapes do, that way the retured reduction 
acts like a regular annotation.

To make things a bit easier I have also locked down the metric to 
`euclidean` for all shapes.

Tests using the symetric flag will come in a later commit.