Update angle average function
The anagle average function now properly finds the angle that minimizes 
the sum((avg_theta - theta)**2) in the bounded interval of accepted 
angles (after accounting for symmetries).

Test cases have been added for using the symmetric flag for the line 
reducers and the new average angle function.

Note: when working with angles the average sometimes is not unique, the 
code might not return the most obvious one, but it does return a 
mathmaticlly-correct one.  E.G. the average of [0, 45, 90] when wrapping 
at 120 (factor=3) is either 45 or 85 (both have the same summed 
distance).  The code will return 85 even though it is the less intuitive 
of the options.