Add GUI for calling argparse scripts
This uses Gooey to create a single GUI that can run all three of the 
command line scirpts.

To have the progress bar propperly printed within the GUI parts of the 
progressbar and gooey class methods need to be overridden, this is done 
in the `gui_overrides` script.

A major version number is being increased to changes in the command line 
calls: the output directory can now be explicitly set for all scripts.