Update gui.py to copy `.exe` file correctly
On windows more files are called in the stack, so the code to get the correct filename to copy is a bit more complex than before.  This code now searches the full call stack for the file with `.exe` in the name and makes a copy of that file without the extension.

If this file is ever updated any widows users will have to manually delete this file before using it (or the `isfile` check should be removed).