Fix poly_line_text_extractor to measure near-vertical slopes better
To account for the cases when multiple dots are drawn per classification 
a line is fitted to the data and the slope measured from this fit.  If 
the line is near-vertial the `polyfit` fails with a `RankWarning` and 
the resulting slope is not accurate.  

In this PR the rank warning has been converted to an error and a `try, 
except` block is used to catch it.  When caught, the points are rotated 
by 90 before being fit.