Add tests and fix warning filter
The tests for both `line` and `word` aggregation now contain vertical 
text (exactly 90 deg) to catch this edge case.  To make the tests pass 
the warning filter had to be adjusted.

The change in the warning filter is inside a `catch_warnings` `with` 
block so the change is scoped to the section of code that needs it.