Add base code for the TESS user reducer
This reducer will calculate the user's skill based on the dificulty of 
the subeject (passed in on `relevant_reduction`) and how many GS 
transits they found (passed in on `data`).  This is a running reducer so 
a `store` is used to keep track of the running `seed` value and the 
number of GS subjects seen (`count`).

For this template I made some reasonable assumptions for the shape of 
the data being passed in and what needs to be returned.