Add custom reducers for TESS (#103)
* Add custom column reducer for TESS

This reducer taks a column extract and returns the location and PDF 
values for all peaks found.  The PDF is constructed from noraml 
distributions centered on the columns drawn.

* Adjust tess reducer to return more values

This reducer now matches all the data returned by Nora's original code.

The PDF, x_eval, and count_ratio values have been added to the list of 
returned values.

* No new packages need to be installed

* Add ability to pass `user_id` and `relevant_rduction` fields into a reducer

These are keywords on the decorator so can be set on a case-by-case 
basis.

* Update test base class to pass keywords on the network request

The `network_kwargs` option is added to allow keywords to be passed 
along the network request and have the reducer wrapper pass them along 
to the reducer.

* Pass new kwargs_extra_data in offline mode using normal keywords

* Update column reducer to use new clustering method

The reducer and test have been updated to reflect the new data 
aggregation method used by TESS.  `user_id` and `relevant_reduction` are 
passed in so the reducer can report what users identified each cluster 
and calculate weighted counts for each cluster.

* Clean up test formatting

* Adds framework for a `running_reducers` route

For these types of reducers a `store` property is passed along that 
contains any information needed to update the returned value based on 
only the most recent extract.

* Add base code for the TESS user reducer

This reducer will calculate the user's skill based on the dificulty of 
the subeject (passed in on `relevant_reduction`) and how many GS 
transits they found (passed in on `data`).  This is a running reducer so 
a `store` is used to keep track of the running `seed` value and the 
number of GS subjects seen (`count`).

For this template I made some reasonable assumptions for the shape of 
the data being passed in and what needs to be returned.

* Update `count` to be the number of transits in a GS subejct

The `count` value should be the total number of transits seed by the 
user, not the number of GS subejcts (there can be multiple transits per 
subject).

* Fix broken import

* Add subject reducer to API

* Update running_reducer_wrapper to pass values in the correct format

Also remove un-needed functionality (i.e. sub-task handeling).

* Add tests for TESS user reducer

The input shape for the extracts, store, and relevant_reduction were 
changed to match how typical extracts are passed in.  See the test case 
for details about the expected structure.

* Change object key from `values` to `value`

* Add doc string to TESS column reducer

* Update documentation to include TESS reducers

* Add reducers to task lookup table doc

* Update TESS reducers to use correct caesar payloads

The Caesar payloads have been finalized and the various reducers and 
wrappers have been updated.

* Add TESS gold standar reducer

This reducer will take the "pluck field extracts" with the GS feedback 
for a subejct.  It returns the diffculty of each transit in the subject.

* Fix relevent reduction strucutre for TESS user reducer

* Add keyword to switch TESS column tool to allow x to be left or center

The old PFE column tool returns the left side of the column, the new PFE 
2.0 tool will return the center of the column.  TESS will have one 
workflow with each type, this keywork will allow both workflows to use 
the same reducer by passing in a `?x=center` or `?x=left` on the 
endpoint.  Defaluts to center.