Update workflow auto-config script for new data export format
The `workflow_contents` csv files are no longer available and the 
`strings` column from it has been moved to the workflow export file.  
This stream-lines much of the configuration code since it only has to 
deal with one csv file.

The documentation and examples have been updated to reflect these 
changes.