Move and test routes.py
The routes.py file has been moved into the application folder to enable 
tests to be run on it.

The new tests check that home, docs, all extractor, all reducer, and all 
running reducer routes return the proper values when a GET request is 
made.

A test POST request is made to a mocked extractor function.

The custom JSON encoder is tested to make sure numpy ints, floats, and 
arrays are converted correctly.