Update scikit-learn to 0.21.1
HDBSCAN now works with the latest scikit-learn but it still has one 
`DepreciationWarning` for its `joblib` import.  I have added a local 
warning filter to catch this and locked down the new scikit-learn 
version.

Scikit-learn needs to be on the new version so the aggergation code can 
have access to its new OPTICS clustering code.