Change tokenizer to split out non-word characters except brackets
There is a bug in the current collatex code that prevents the alignment 
from working if more than six witnesses contain `'$'` without spaces 
after them.  This is a side effect of tokenizing on whitespace instead 
of the defulat of splitting on non-word characters.

The tokenization regex has been updated so non-word characters 
(excluding `[]<>{}`) are split off on their own.

The tests have been updated to match this now tokenization.