Update extractor_wrapper to optionally track `gold_standard`
When the wrapper is used with `gold_standard=True` the `gold_standard` 
key from an annotaiton object will be passed into the extracton fuction 
with the `gold_standard` keyword.  The extractor decides how this should 
be strcutrued inside the extract.

Since the extractor_wrapper takes a keyword all the 
`@extractor_wrapper`s had to be changed to `@extractor_wrapper()`.