Add running reducer for TESS gold standard reducer
This will help Caesar process the TESS backlog faster by switching all reducers over to running mode.

In an effort to re-use code the `process_data` function is imported from the existing non-running reducer of the same name.

To avoid errors in Flask the underlying function for this reducer needs a *unique* name, this is why `_rr` has been added at the end of the function name.  The `__init__.py` file in the `running_reducer` folder removes this `_rr` when defining the route name, so the API endpoint will be `/running_reducers/tess_gold_standard_reducer`.