Add a running reducer for the TESS column reducer
This will help caesar avoid the issue when a workflow has "mixed type" reducer (i.e. some "default" and some "running").

The running reducer wrapper has had the `user_id` and `relevant_reduction` keywords added to match the normal reducer wrapper.

The original reducer is called to do the majoity of the work for this running reducer (accessed via the `._original` name so the wrapper is bypassed), the new code handels using the results of the `process_data` function to update the store. Both the `user_id` and `relevant_reduction` lists are also tracked in the store.

When the store is passed back into the function via the mock flaks test the list-of-tuples produced by `process_data` was converted to a list-of-lists. As the data type does not matter in this case I updated `process_data` to produce a list-of-lists in the first place.