Optimize question reducer
The question reducer used `Counter` as an intermediate container for the data. It turns out that summing these `Counter`s is quite slow.

To speed up the reducer a list is made of all answers given (across all extracts) and `Counter` is called at the end to get a count for all unique answers.  This new code is 60% to 85% faster than the previous method (more speed up for larger extracts).