Add mulit-core reducing
This splits reductions across multiple cores (if available).  Each subject is sent to a different porcess, the results will be written in the order they are returned (i.e. random-ish).

The tests have been updated to not check reducer mocks driectly as the ID changes when they are pickled for multiprocessing.

A new command line option `-c` is added to adjust the number of cores used.