Ensure small ammounts of data are passed through `apply_async`
This data needs to be pickled and sent to each process, and is the main IO bottleneck for multiprocessing.  By only passing along the needed rows of extracts to the `reduce_subejcet` function all multiprocessing goes faster.

Each task is now queued up a seperate call to the `reduce_subject` fucntion instead of one call per subject.  In streaming mode this means more wirtes are made to disk.  The tests have been updated to account for this.