Add consensus text for each line
Before only the alignment table was provided, now a new key `consensus_text` is available that contains the most common word from the rows of the aligment table.

This has been bundled into the `consensus_score` sub-function that is shared by all of the text reducers.  It also ensures that this is the same words used to calculated the `consensus_score` are in this text.  Words are joined with a single space.

All tests have been updated.