Add parameters to optics reduction
The ALICE front-end needs to know what parameters the text redcuers run with.

For this update the parameters are added in the `reducer_wrapper` to ensure all keywords are added after defaults are added in.  This change requried the `base_test_class` to be updated so the parmeters are removed when testing the unwrapped function.