Small update to csv_utils and routes
The JSON encoder will now convert `np.bool_` to `bool` just in case it becomes an issue in the future.  this is a case that happens more often in the offline code when passing things through pandas.

The `unflatten_data` function has been updated to use `eval` direclty so it can also define `nan = None`.