Add subject reducer for Gravity Spy
This reducer takes the confusion matrix of a user and a ML weight vector to find the probability a subject belongs to the voted catagories.  This is written as a running reducer as this is the only means of retirement on the project so one subject could have a large number of extracts.  Plus the confusion matrix for each extract is needed as well.

If the column of a confusion matrix is all zeros or there is not confusion matrix at all this will return the current approximation (act as if they do not exist).