Track "none of the above" answers
When "none of the above" is voted for increment 'none_of_the_above_count'.  This should not cause the 'number_views' to increase, or change any other value in the store.

This value will be used to move subjects between levels in the project.

The `none_key` can be set as a parameter of the reducer (default "NONEOFTHEABOVE").