Fix `isfinite` bug and change `max_eps` default
The `isfinite` but was caused by code that should have been removed, but still stuck around in the `cluster_of_one` sub-fuction.  Not logged in users are passed along as `None` (even when converted to a numpy array), not `np.nan`.

The old default for `max_eps` as `np.inf` but this does not pass along in JSON very well, so instead the new default is `None` and the optics reducer converts this to `np.inf` when run.