Working extractor sub_task wrapper
The subtask wrapper checks the classifier version and when it is `2.0` or higher is usees the new format for subtasks.

First it splits the annotation into the drawing task annotation and the subtask annotations by looking at the `taskType` key.

The drawing annotation is passed to the shape extractor as normal.  After that is done the subtasks are extracted if a extractor is defined in the `details` url param (see test for format).

The subtask key and mark order from the drawing task are used to find and order the subtasks.  Once extracted they are placed in the extract with the `T*_tool*_subtask*` key as a list in the order as the marks.

*To Do* allow the `task` url param to pass in all the subtasks in addition to the base drawing task.