Impliment v2.0 subtask reduction
This completes the work on the v2.0 subtask reducer.  With the new extract format the resulting code is much easier to read and the resulting reduction is cleaner.  This structure will flatten nicely with one new column added for each subtask.

The url params match the ones passed into the extractor.