Pass user_id to subtasks
This requires that user id be pased to the base shape reducer as well. This change resulted in all the shape tests needing to be updated to pass mock user id data along all the time.

This also resulted in more of the reducer test factroies needing to be updated to handle this extra test data.

In both online and offline mode all reducers are passed user id data already, so there is no need to change any of the base API code, just needed to update the reducer wrapper to pass this info along to the reducer function.