Fix #388
Add documentation for `minimum_views` value on the transription reducers.  This value is passed along to the fron-end and has no effect on the reduction results.  I think any further documentation for how the front-end uses this value should belong with the front-end code so can stay up to date if it ever changes.

Happy to link to the documenation from here if a link exists.