bump numpy to latest version
a bug in pip means it builds other packages against numpy 1.20.0 regardless of what verion of numpy is pinned!  This forces the code to use the latest numpy to work at all!

This version of numpy was being avoided because it broke one of the tests.  Time to fix that bug...