Fix bug with latest numpy
The numpy issue is a line with slope exactly equal to pi was getting a `dy` of `-0` rather than `0`.  Since `-0` is distict from `0` in `rad2deg` the slope was coming back as `-pi` rather than `pi`.

The fix is to check for `-0` and cahnge it to `0` (see code changes).  In numpy `-0 == 0` is `True` so that helps.

This is ultimitly a change in the floting point rounding error intoduced with the new numpy.

For reference numpy 1.20.0 change the version of `C` it was compiled against, so this changes things like the floting point errors.