Adjust how userify is tested when in offline mode
In offline mode userfy should not be tested.  This PR adjusts the unit tests to check if the package was installed on online mode before running any tests.  I also needed to adjust how the __init__ file handeled a failed import.

This was tested in the local scripts docker container that installs the code on offline mode.