Added new average, based on n polygons overlapping
A new average for the polygon cluster has been added, based on the intersection/overlap of n polygons. The idea is to find the region where at least n volumteers agree. An algorithm was developed for this purpose, details of which are given in the doc string of this new function. A more complete and general function to find the union of intersecting polygons was also added.