Update optics test
This makes sure that the `xi` is not set when testing the reducer in full to ensure that the new default value is being passed in correctly.

To keep the tests working, this keyword needs to be included in the `okwargs` section, as these are the keywords passed in to the "original" function before the wrapper that sets the new defaults.