Update docker containers to use the latest version of python
For the server to work on python 3.13 and use `panoptes-client` two new dependencies needed to be added, `setuptools` and `standard-imghdr` as these are no longer included in Python.

Ideally these should be in the client's dependency file, but for now this works.

Using some carful version lower bounds we are able to keep python 3.9 support with all the latest code changes.  There is only one numpy import that is in a `try` block to make the code work with either version of numpy.

Finally, due to some odd things with folder mounting into the docker container on my mac, I am truing off the code folder mounting for the server container.  I am not sure if this is just local folder permissions, or some new change to docker, either way it is likely better to be on the safe side with the container that is deployed.