Update the github actions
For unittests expand the tested python versions out to 3.13 and use the inbuilt `cache` command for the python containers (no need to do this by hand anymore :D)

On the PyPi actions, update them to use `flit publish` directly with environment variables.  This is cleaner than depending on a 3rd-party action that in the end can be done with one line of code.