Removed TestReduceSurvey
The survey reducer used to have its own, unique test factory.This is because it was the only reducer to return a list rather than a dictionary. As the base reducer tests can now handle lists of dictionaries, this became redundent. To facilitate its removal the survey reducer and its unit test were updated. Main change was 'vote_count' is now a part of the processed data rather than being its own argument. A function to order the list alphbetically in terms of 'choice' was also added