Ensure progressbar prints correctly in the GUI
progressbar2 added support for colors in the terminal, by default the "percent" goes from yellow to green.  Gooey does not support this, and as a result it looks confusing.

The override code I use ended up stripping out the checks for the flags that would normally turn this off, so I updated the override to force it off myself.