Flake8 update
The latest Flake8 updated has started tagging lines that were fine before.  Specifically all the `nonlocal` variables in `batch_utils.py` are being flagged as "not defined in scope" while this is true, this is also how `nonlocal` works in the first place...  I have added `# noqa: F824` to each use of `nonlocal` to suppress the false flag error.