Add the ability for fast IoU average estimates
A new keyword for all reducers that support IoU clustering that defaults to previous behavior.  The new estimate takes the "most typical" reduction and returns that as the average.  This is far faster than calculating the true average and can bring reducer run times for a typical project down from several hours to a few mins.